Move string functions into string module
